New example: Dynamic Range Demo
Added a new example JSON demonstrating a runnable model that passes min, max, and step values
to a range, to modify the values the user can input to the simulation.

Also updated the Dynamic Selector demo with Gauge-related addon info from the Authoring Tool.
For the purposes of the JSON Schema repo, this may serve as a good addon usage example.